// packages/server/src/Interface.ts
export enum Platform {
    OPEN_SOURCE = 'open source',
    CLOUD = 'cloud',
    IAM = 'iam'
}